Factory method para leer los .tmx, el bug del minijuego GoingUp aún sigue, pero la carga del mapa se ha optimizado aún más
Recompilad :blush:

+++ Mapa/Nodo/Ground.cpp
@@ -13,7 +13,7 @@
 
 #include "Ground.h"
 
-Ground::Ground(std::vector<std::array<float,2> > coords) {
+Ground::Ground(std::vector<std::array<int,2> > coords) {
     physicsEngine* world;
     
     t = new physicsEngine::type;
